Initial working version.
This code works. It autoboots the cartridge and starts the emulator with BANK1 pre-loaded. Press "1" to start emulation. Press <SHIFT><COMMODORE> to get back to menu. BUG: CBM-II cursor remains on screen. To test in VICE load the 3 binaries into the rom slots and reset CPU. To test on a real machine you need to burn the binaries to eprom and install in 24K cartridge board, or use my CBM-II ROMCART.